requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth
beautifulsoup4
requests
openai
python-dotenv
